open-amp: Add default SRCREV to open-amp-dev.bb
open-amp-dev.bb is supposed to always uses latest of master branch
if open-amp-dev is specified as preferred provider.

In order to allow no network mode, we add a default SRCREV not to
use auto if open-amp-dev is not specified as preferred provider.
So that we yocto pass all the recipes, it will not try to access
the network.

Signed-off-by: Wendy Liang <jliang@xilinx.com>