rpmsg-prox-module: use local ept in open() to send msg
In 4.14 kernel, when probing rpmsg device, it will not
create endpoint if no default callback is provided by the rpmsg
driver. In open() implementation of the proxy module, use the
created proxy endpoint to send message instead of using the
default rpmsg device endpoint as there is no default rpmsg endpoint
in this case.

Signed-off-by: Wendy Liang <jliang@xilinx.com>