yocto-cmake-translation: Adjust table to current system
Issue: CR-1099768

By using 'MACHINE' in the previous code, it made any users of this class
MACHINE specific packages, even if MACHINE wasn't actually consulted for
their particular use-cases.

Further it was determined the MACHINE check was no longer valid due to
much of the changing in the current versions of the code.  Refactored
the table based on discussions with the open-amp folks.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>