open-amp: decoupling: change SDT variable to SYSTEM_DTFILE_OPENAMP
As Linux device tree decoupling recipes has different expectations on which
device tree is present, work around this for now so that the variable between
both is not common. The original variable SYSTEM_DTFILE is overwritten
before open-amp uses the variable and this unblocks the open-amp decoupling
flow for linux.

Signed-off-by: Ben Levinsky <ben.levinsky@xilinx.com>
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>