rpmsg-examples: Change endpoint destination to 0
Previously, when the host attempted to create its RPMsg endpoint, the
destination was set to RPMSG_ADDR_ANY. Change this to 0 to match the remote
firmware's corresponding endpoint source of 0.

This change is required as the kernel no longer infers and sets the
host-endpoint's corresponding channel and instead requires the following to
be specified:
- name   (host already does this)
- source (host already does this)
- dest   (host DID NOT do this)

The dest field specified in the host endpoint is now set to the firmware's
source field.

With this change, the kernel will locate the endpoint and corresponding channel
that was created when the RPMsg-compatible firmware was started beforehand and
the demo will proceed as before.

This has been tested with both 1 and 2 concurrent remotes talking to host to
ensure this fix scales to multiple remotes talking in concurrent fashion.

Signed-off-by: Ben Levinsky <ben.levinsky@xilinx.com>
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>