layer.conf: Rework vendor specific inclusions
Including the vendor layer information using ${@...} syntax is not reliable
as the order of the BBFILES_COLLECTIONS can change depending on layer.conf
and other factors.

Instead move to include all of the vendor layer.inc files, but set the rule
that a layer.inc file must only use BBFILES_DYNAMIC and any other settings
must not conflict with non-vendor or other potential vendor's setitngs.

Signed-off-by: Mark Hatle <mark.hatle@kernel.crashing.org>