Add new recipe to add remoteproc nodes in device-tree
remoteproc nodes are not merged in upstream kernel. We are maintaining
openamp.dtsi separately to include remoteproc related nodes.

This change includes remoteproc nodes via openamp.dtsi for all zynqmp
boards.

Signed-off-by: Tanmay Shah <tanmay.shah@xilinx.com>

This change is specific to device-tree output from meta-xilinx-tools'
DTG generation.  So only apply if meta-xilinx-tools is available.

Reworked to check that system-top.dts is available, and only update it
if it is.

Also move the path and SRC_URI from using zynqmp override, as it can cause
recipetool issues.

Signed-off-by: Mark Hatle <mhatle@xilinx.com>

Reformat to work with latest meta-openamp

Signed-off-by: Ben Levinsky <ben.levinsky@xilinx.com>