linux: move UIO enable to BSP specific config file.
UIO is a controversial feature.  We should not enable it in the openamp.cfg
which is used by every kernel and every machine (when openamp is in
DISTRO_FEATURES).

Instead move it to remoteproc-*.cfg as that is only included by
specific machines (and also only when enabled via openamp DISTRO_FEATURE).

The old config was not doing any good anyway as the base level UIO
was not enabled.  We fix this here also.

Signed-off-by: Bill Mills <bill.mills@linaro.org>